WiP: refactor ingest to use SQS
