+++ receiver.tf
@@ -1,7 +1,7 @@
 data "archive_file" "receiver_zip" {
   type        = "zip"
-  source_file = "${path.module}/receiver/receiver.js"
-  output_path = "${path.module}/receiver/receiver.zip"
+  source_file = "${path.module}/lambdas/receiver.js"
+  output_path = "${path.module}/lambdas/receiver.zip"
 }
 
 resource "aws_lambda_function" "receiver_lambda" {
@@ -15,7 +15,8 @@ resource "aws_lambda_function" "receiver_lambda" {
 
   environment {
     variables = {
-      S3_BUCKET = aws_s3_bucket.receiver.bucket
+      SQS_QUEUE_URL = aws_sqs_queue.receiver_queue.id
+      AWS_REGION = var.aws_region
     }
   }
 }
@@ -93,4 +94,8 @@ resource "aws_api_gateway_deployment" "deployment" {
   variables = {
     "S3_BUCKET" = aws_s3_bucket.receiver.bucket
   }
+}
+
+resource "aws_sqs_queue" "receiver_queue" {
+  name = "controlshift-received-webhooks"
 }
